backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
copilotkit==0.1.18
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
idna==3.10
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langgraph==0.2.34
langgraph-checkpoint==2.0.0
langgraph-sdk==0.1.32
langsmith==0.1.130
marshmallow==3.22.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.51.0
orjson==3.10.7
packaging==24.1
partialjson==0.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
